Replace PathSegmentExtractor with TemplateTokenizer
- As fine-grained errors don't matter on structural mismatch
- We can use a regex based approach which simplifies the extraction
- Move tests from PathSegmentExtractorTest to HttpPathPatternTest
