+++ core/src/main/kotlin/io/specmatic/core/HttpPathPattern.kt
@@ -6,6 +6,8 @@ import io.specmatic.conversions.convertPathParameterStyle
 import io.specmatic.core.Result.Failure
 import io.specmatic.core.Result.Success
 import io.specmatic.core.pattern.*
+import io.specmatic.core.utilities.ensurePrefix
+import io.specmatic.core.utilities.ensureSuffix
 import io.specmatic.core.value.StringValue
 import io.specmatic.core.value.Value
 import java.net.URI
@@ -17,7 +19,7 @@ data class HttpPathPattern(
     private val path: String,
     private val otherPathPatterns: Collection<HttpPathPattern> = emptyList(),
 ) {
-    private val pathSegmentExtractor: PathSegmentExtractor = PathSegmentExtractor(path, pathSegmentPatterns)
+    private val pathSegmentExtractor: TemplateTokenizer = createTokenizerFromPathSegments()
     private fun calculateSpecificity(): Int {
         val segments = this.path.removePrefix("/").removeSuffix("/").split("/")
         return segments.sumOf(::segmentSpecificity)
@@ -81,7 +83,7 @@ data class HttpPathPattern(
 
         val pathSegments = extractPathSegments(path)
         if (pathSegmentPatterns.size != pathSegments.size) {
-            return Failure("Expected $path to match ${this.path}.", breadCrumb = BreadCrumb.PATH.value, failureReason = FailureReason.URLPathMisMatch)
+            return Failure("Failed to extract segments for ${this.path} from $path", breadCrumb = BreadCrumb.PATH.value, failureReason = FailureReason.URLPathMisMatch)
         }
 
         val results = checkIfPathSegmentsMatch(pathSegments, resolver, path)
@@ -345,6 +347,14 @@ data class HttpPathPattern(
         }
     }
 
+    fun List<String>.joinToPath(): String {
+        return ensurePrefixAndSuffix(joinToString(separator = ""))
+    }
+
+    fun ensurePrefixAndSuffix(rawPath: String): String {
+        return rawPath.normalizePrefix(path).normalizeSuffix(path).normalizeSlash()
+    }
+
     fun extractPathParams(requestPath: String, resolver: Resolver): Map<String, Value> {
         val pathSegments = extractPathSegments(requestPath)
         return pathSegmentPatterns.zip(pathSegments).mapNotNull { (pattern, value) ->
@@ -391,7 +401,7 @@ data class HttpPathPattern(
         }
 
         if (pathSegmentPatterns.none { it.key == parameterName }) return null
-        return updatedSegments.joinToString("")
+        return updatedSegments.joinToPath()
     }
 
     fun zipWithExpandedPathSegments(other: HttpPathPattern, resolver: Resolver): List<Pair<URLPathSegmentPattern, URLPathSegmentPattern>>? {
@@ -438,9 +448,13 @@ data class HttpPathPattern(
         }
     }
 
+    fun extractPathSegments(rawPath: String): List<String> {
+        return pathSegmentExtractor.extract(ensurePrefixAndSuffix(rawPath))
+    }
+
     private fun expandPath(other: HttpPathPattern, resolver: Resolver): HttpPathPattern {
         if (this.pathSegmentPatterns.size == other.pathSegmentPatterns.size) return other
-        val segments = pathSegmentExtractor.extract(other.path)
+        val segments = extractPathSegments(other.path)
         if (segments.size != this.pathSegmentPatterns.size) return other
         return other.copy(pathSegmentPatterns = pathSegmentPatterns.zip(segments).map { (pattern, segment) ->
             if (pattern.pattern is ExactValuePattern) return@map URLPathSegmentPattern(ExactValuePattern(StringValue(segment)))
@@ -468,27 +482,40 @@ data class HttpPathPattern(
         return true
     }
 
-    private fun extractPathSegments(rawPath: String): List<String> {
-        return pathSegmentExtractor.extract(rawPath)
-    }
-
     private fun splitPathBySlash(path: String): List<String> {
         val trimmed = path.trim('/')
         if (trimmed.isEmpty()) return emptyList()
         return trimmed.split("/").filter(String::isNotBlank)
     }
 
-    private fun List<String>.joinToPath(): String {
-        val combined = joinToString(separator = "")
-        return pathSegmentExtractor.ensurePrefixAndSuffix(combined)
+    private fun String.normalizePrefix(contractPath: String): String {
+        return if (contractPath.startsWith('/')) ensurePrefix("/") else removePrefix("/")
+    }
+
+    private fun String.normalizeSuffix(contractPath: String): String {
+        if (contractPath == "/") return this
+        return if (contractPath.endsWith('/')) ensureSuffix("/") else removeSuffix("/")
+    }
+
+    private fun String.normalizeSlash(): String = replace(Regex("/{2,}"), "/")
+
+    private fun createTokenizerFromPathSegments(): TemplateTokenizer {
+        val regexSource = pathSegmentPatterns.joinToString(prefix = "^", separator = "", postfix = "$") { pattern ->
+            when (val p = pattern.pattern) {
+                is ExactValuePattern -> Regex.escape(p.pattern.toUnformattedString())
+                else -> NON_SLASH_REGEX
+            }
+        }
+        return TemplateTokenizer(Regex(regexSource))
     }
 
     companion object {
+        private const val NON_SLASH_REGEX = "([^/]+)"
         internal val internalPathRegex: Regex = Regex("\\([^():]+:[^()]+\\)")
 
-        internal fun extractUsingInternalPathRegex(path: String): TemplateTokenizer {
+        internal fun createTokenizerFromInternalPathRegex(path: String): TemplateTokenizer {
             val parts = internalPathRegex.split(path)
-            val pattern = parts.joinToString(separator = "([^/]+)") { Regex.escape(it) }
+            val pattern = parts.joinToString(separator = NON_SLASH_REGEX) { Regex.escape(it) }
             return TemplateTokenizer(Regex(pattern))
         }
 
@@ -512,7 +539,7 @@ internal fun buildHttpPathPattern(
 }
 
 internal fun pathToPattern(rawPath: String): List<URLPathSegmentPattern> {
-    val segments = HttpPathPattern.extractUsingInternalPathRegex(rawPath).extract(rawPath)
+    val segments = HttpPathPattern.createTokenizerFromInternalPathRegex(rawPath).extract(rawPath)
     return segments.map { part ->
         when {
             isPatternToken(part) -> {
