+++ core/src/test/kotlin/io/specmatic/core/HttpPathPatternTest.kt
@@ -15,7 +15,9 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.CsvSource
+import org.junit.jupiter.params.provider.MethodSource
 import java.io.UnsupportedEncodingException
 import java.net.URI
 import java.net.URISyntaxException
@@ -157,8 +159,8 @@ internal class HttpPathPatternTest {
             assertThat(it is Result.Failure).isTrue()
             assertThat((it as Result.Failure).toMatchFailureDetails()).isEqualTo(
                 MatchFailureDetails(
-                    listOf("PARAMETERS.PATH (/owners/123123)"),
-                    listOf("""Expected "/pets/", actual was "/owners/"""")
+                    listOf("PATH"),
+                    listOf("""Failed to extract segments for /pets/(petid:number) from /owners/123123""")
                 )
             )
         }
@@ -913,5 +915,144 @@ internal class HttpPathPatternTest {
             val updated = pathPattern.updatePathParameter("/test/first/second/status", "id1", StringValue("updated"))
             assertThat(updated).isEqualTo("/test/updated/second/status")
         }
+    }
+
+    @Nested
+    inner class NormalizationTests {
+        @ParameterizedTest(name = "contract \"{0}\", input \"{1}\" -> \"{2}\"")
+        @MethodSource("io.specmatic.core.HttpPathPatternTest#normalizationCases")
+        fun `ensurePrefixAndSuffix should normalize based on contract path`(contractPath: String, rawPath: String, expected: String) {
+            val normalized = HttpPathPattern.from(contractPath).ensurePrefixAndSuffix(rawPath)
+            assertThat(normalized).isEqualTo(expected)
+        }
+    }
+
+    @Nested
+    inner class ExtractionTests {
+        @Test
+        fun `should extract pure literals deterministically`() {
+            val extractor = HttpPathPattern.from("/pets/list")
+            val segments = extractor.extractPathSegments("pets/list")
+            assertThat(segments).containsExactly("/pets/list")
+        }
+
+        @Test
+        fun `should extract literal variable literal variable path`() {
+            val extractor = HttpPathPattern.from("/pets/(id:string)/owners/(owner:string)")
+            val segments = extractor.extractPathSegments("/pets/123/owners/alice")
+            assertThat(segments).containsExactly("/pets/", "123", "/owners/", "alice")
+        }
+
+        @Test
+        fun `should extract when variable appears at start and end`() {
+            val extractor = HttpPathPattern.from("/(tenant:string)/orders/(orderId:string)")
+            val segments = extractor.extractPathSegments("/acme/orders/42")
+            assertThat(segments).containsExactly("/", "acme", "/orders/", "42")
+        }
+
+        @Test
+        fun `should honor trailing slash normalization for extraction`() {
+            val extractor = HttpPathPattern.from("/pets/(id:string)/")
+            val segments = extractor.extractPathSegments("/pets/123")
+            assertThat(segments).containsExactly("/pets/", "123", "/")
+        }
+
+        @Test
+        fun `should extract slash separated continuous parameters`() {
+            val extractor = HttpPathPattern.from("/test/(id1:string)/(id2:string)/status")
+            val segments = extractor.extractPathSegments("/test/first/second/status")
+            assertThat(segments).containsExactly("/test/", "first", "/", "second", "/status")
+        }
+    }
+
+    @Nested
+    inner class InterpolatedExtractionTests {
+        @Test
+        fun `should extract hyphen interpolated path`() {
+            val extractor = HttpPathPattern.from("/product/product-(id:string)/order/order-(orderId:string)/latest")
+            val segments = extractor.extractPathSegments("/product/product-12/order/order-abc/latest")
+            assertThat(segments).containsExactly("/product/product-", "12", "/order/order-", "abc", "/latest")
+        }
+
+        @Test
+        fun `should extract comma separated interpolated path`() {
+            val extractor = HttpPathPattern.from("/test/(id1:string),(id2:string)/status")
+            val segments = extractor.extractPathSegments("/test/first,second/status")
+            assertThat(segments).containsExactly("/test/", "first", ",", "second", "/status")
+        }
+
+        @Test
+        fun `should extract mixed literals around variables in the same segment`() {
+            val extractor = HttpPathPattern.from("/x-(a:string)-y-(b:string).json")
+            val segments = extractor.extractPathSegments("/x-10-y-20.json")
+            assertThat(segments).containsExactly("/x-", "10", "-y-", "20", ".json")
+        }
+
+        @Test
+        fun `should extract repeated separator boundaries`() {
+            val extractor = HttpPathPattern.from("/r--(left:string)--(right:string)--end")
+            val segments = extractor.extractPathSegments("/r--A--B--end")
+            assertThat(segments).containsExactly("/r--", "A", "--", "B", "--end")
+        }
+    }
+
+    @Nested
+    inner class IntegrationParityTests {
+        @Test
+        fun `should extract with interpolated hyphen path and match successfully`() {
+            val contractPath = "/product/product-(id:string)/order/order-(orderId:string)/latest"
+            val rawPath = "/product/product-12/order/order-abc/latest"
+            val pathPattern = HttpPathPattern.from(contractPath)
+            val segments = pathPattern.extractPathSegments(rawPath)
+            assertThat(segments).containsExactly("/product/product-", "12", "/order/order-", "abc", "/latest")
+            assertThat(pathPattern.matches(URI(rawPath), Resolver())).isInstanceOf(Result.Success::class.java)
+            assertThat(pathPattern.extractPathParams(rawPath, Resolver()).toStringMap()).isEqualTo(mapOf("id" to "12", "orderId" to "abc"))
+        }
+
+        @Test
+        fun `should extract comma separated interpolated params`() {
+            val contractPath = "/test/(id1:string),(id2:string)/status"
+            val rawPath = "/test/first,second/status"
+            val pathPattern = HttpPathPattern.from(contractPath)
+            val segments = pathPattern.extractPathSegments(rawPath)
+            assertThat(segments.joinToString("")).isEqualTo(pathPattern.ensurePrefixAndSuffix(rawPath))
+            assertThat(pathPattern.matches(URI(rawPath), Resolver())).isInstanceOf(Result.Success::class.java)
+            assertThat(pathPattern.extractPathParams(rawPath, Resolver()).toStringMap()).isEqualTo(mapOf("id1" to "first", "id2" to "second"))
+        }
+
+        @Test
+        fun `should extract slash separated interpolated params`() {
+            val contractPath = "/test/(id1:string)/(id2:string)/status"
+            val rawPath = "/test/first/second/status"
+            val pathPattern = HttpPathPattern.from(contractPath)
+            val segments = pathPattern.extractPathSegments(rawPath)
+            assertThat(segments.joinToString("")).isEqualTo(pathPattern.ensurePrefixAndSuffix(rawPath))
+            assertThat(pathPattern.matches(URI(rawPath), Resolver())).isInstanceOf(Result.Success::class.java)
+            assertThat(pathPattern.extractPathParams(rawPath, Resolver()).toStringMap()).isEqualTo(mapOf("id1" to "first", "id2" to "second"))
+        }
+
+        @Test
+        fun `should report HttpPathPattern mismatch when structural mismatch occurs`() {
+            val contractPath = "/product/product-(id:string)/order/order-(orderId:string)/latest"
+            val rawPath = "/product/12/order/abc/latest"
+            val pathPattern = HttpPathPattern.from(contractPath)
+            val result = pathPattern.matches(URI(rawPath), Resolver())
+            assertThat(result).isInstanceOf(Result.Failure::class.java)
+            assertThat((result as Result.Failure).failureReason).isEqualTo(FailureReason.URLPathMisMatch)
+        }
+    }
+
+    companion object {
+        @JvmStatic
+        fun normalizationCases(): List<Arguments> = listOf(
+            Arguments.of("/pets/list", "pets/list", "/pets/list"),
+            Arguments.of("pets/list", "/pets/list", "pets/list"),
+            Arguments.of("/pets/list/", "/pets/list", "/pets/list/"),
+            Arguments.of("/pets/list", "/pets/list/", "/pets/list"),
+            Arguments.of("/pets/list", "///pets//list", "/pets/list"),
+            Arguments.of("/pets/list/", "pets///list", "/pets/list/"),
+            Arguments.of("/", "", "/"),
+            Arguments.of("/", "orders", "/orders")
+        )
     }
 }
